Refactor requests and responses from model to action
