+++ opensearch-observability/src/main/kotlin/org/opensearch/observability/action/CreateObservabilityObjectRequest.kt
@@ -3,12 +3,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.observability.model
+package org.opensearch.observability.action
 
 import org.opensearch.action.ActionRequest
 import org.opensearch.action.ActionRequestValidationException
-import org.opensearch.action.ValidateActions
-import org.opensearch.common.Strings
 import org.opensearch.common.io.stream.StreamInput
 import org.opensearch.common.io.stream.StreamOutput
 import org.opensearch.common.io.stream.Writeable
@@ -19,24 +17,27 @@ import org.opensearch.common.xcontent.XContentParser
 import org.opensearch.common.xcontent.XContentParserUtils
 import org.opensearch.commons.utils.fieldIfNotNull
 import org.opensearch.commons.utils.logger
+import org.opensearch.observability.model.BaseObjectData
+import org.opensearch.observability.model.ObservabilityObjectDataProperties
+import org.opensearch.observability.model.ObservabilityObjectType
 import org.opensearch.observability.model.RestTag.OBJECT_ID_FIELD
 import java.io.IOException
 
 /**
  * Action request for creating new configuration.
  */
-internal class UpdateObservabilityObjectRequest : ActionRequest, ToXContentObject {
-    val objectId: String
+internal class CreateObservabilityObjectRequest : ActionRequest, ToXContentObject {
+    val objectId: String?
     val type: ObservabilityObjectType
     val objectData: BaseObjectData?
 
     companion object {
-        private val log by logger(UpdateObservabilityObjectRequest::class.java)
+        private val log by logger(CreateObservabilityObjectRequest::class.java)
 
         /**
          * reader to create instance of class from writable.
          */
-        val reader = Writeable.Reader { UpdateObservabilityObjectRequest(it) }
+        val reader = Writeable.Reader { CreateObservabilityObjectRequest(it) }
 
         /**
          * Creator used in REST communication.
@@ -45,7 +46,7 @@ internal class UpdateObservabilityObjectRequest : ActionRequest, ToXContentObjec
          */
         @JvmStatic
         @Throws(IOException::class)
-        fun parse(parser: XContentParser, id: String? = null): UpdateObservabilityObjectRequest {
+        fun parse(parser: XContentParser, id: String? = null): CreateObservabilityObjectRequest {
             var objectId: String? = id
             var type: ObservabilityObjectType? = null
             var baseObjectData: BaseObjectData? = null
@@ -73,10 +74,9 @@ internal class UpdateObservabilityObjectRequest : ActionRequest, ToXContentObjec
                     }
                 }
             }
-            objectId ?: throw IllegalArgumentException("$OBJECT_ID_FIELD field absent")
             type ?: throw IllegalArgumentException("Object data field absent")
             baseObjectData ?: throw IllegalArgumentException("Object data field absent")
-            return UpdateObservabilityObjectRequest(baseObjectData, type, objectId)
+            return CreateObservabilityObjectRequest(objectId, type, baseObjectData)
         }
     }
 
@@ -93,13 +93,14 @@ internal class UpdateObservabilityObjectRequest : ActionRequest, ToXContentObjec
 
     /**
      * constructor for creating the class
-     * @param objectData the ObservabilityObject
      * @param objectId optional id to use for ObservabilityObject
+     * @param type type of ObservabilityObject
+     * @param objectData the ObservabilityObject
      */
-    constructor(objectData: BaseObjectData, type: ObservabilityObjectType, objectId: String) {
-        this.objectData = objectData
-        this.type = type
+    constructor(objectId: String? = null, type: ObservabilityObjectType, objectData: BaseObjectData) {
         this.objectId = objectId
+        this.type = type
+        this.objectData = objectData
     }
 
     /**
@@ -107,7 +108,7 @@ internal class UpdateObservabilityObjectRequest : ActionRequest, ToXContentObjec
      */
     @Throws(IOException::class)
     constructor(input: StreamInput) : super(input) {
-        objectId = input.readString()
+        objectId = input.readOptionalString()
         type = input.readEnum(ObservabilityObjectType::class.java)
         objectData = input.readOptionalWriteable(
             ObservabilityObjectDataProperties.getReaderForObjectType(
@@ -124,7 +125,7 @@ internal class UpdateObservabilityObjectRequest : ActionRequest, ToXContentObjec
     @Throws(IOException::class)
     override fun writeTo(output: StreamOutput) {
         super.writeTo(output)
-        output.writeString(objectId)
+        output.writeOptionalString(objectId)
         output.writeEnum(type)
         output.writeOptionalWriteable(objectData)
     }
@@ -133,10 +134,6 @@ internal class UpdateObservabilityObjectRequest : ActionRequest, ToXContentObjec
      * {@inheritDoc}
      */
     override fun validate(): ActionRequestValidationException? {
-        var validationException: ActionRequestValidationException? = null
-        if (Strings.isNullOrEmpty(objectId)) {
-            validationException = ValidateActions.addValidationError("objectId is null or empty", validationException)
-        }
-        return validationException
+        return null
     }
 }
