+++ opensearch-observability/src/main/kotlin/org/opensearch/observability/action/GetObservabilityObjectResponse.kt
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.observability.model
+package org.opensearch.observability.action
 
 import org.opensearch.common.io.stream.StreamInput
 import org.opensearch.common.io.stream.StreamOutput
@@ -11,6 +11,9 @@ import org.opensearch.common.io.stream.Writeable
 import org.opensearch.common.xcontent.ToXContent
 import org.opensearch.common.xcontent.XContentBuilder
 import org.opensearch.common.xcontent.XContentParser
+import org.opensearch.observability.model.BaseResponse
+import org.opensearch.observability.model.ObservabilityObjectSearchResult
+import org.opensearch.observability.model.RestTag
 import java.io.IOException
 
 /**
