+++ opensearch-observability/src/main/kotlin/org/opensearch/observability/action/UpdateObservabilityObjectResponse.kt
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.observability.model
+package org.opensearch.observability.action
 
 import org.opensearch.common.io.stream.StreamInput
 import org.opensearch.common.io.stream.StreamOutput
@@ -13,6 +13,7 @@ import org.opensearch.common.xcontent.XContentParser
 import org.opensearch.common.xcontent.XContentParser.Token
 import org.opensearch.common.xcontent.XContentParserUtils
 import org.opensearch.observability.ObservabilityPlugin.Companion.LOG_PREFIX
+import org.opensearch.observability.model.BaseResponse
 import org.opensearch.observability.model.RestTag.OBJECT_ID_FIELD
 import org.opensearch.observability.util.logger
 import java.io.IOException
